Moved constants.py to src/rules-engine
